Removed custom WrappingCanvas control for now, replaced with VariableSizedWrapGrid. Updated EmoteControl to animate and updated view.

+++ Emotible/Emotible/ViewModels/ISizableControl.cs
@@ -8,7 +8,7 @@ namespace Emotible.ViewModels
 {
     public interface ISizableControl
     {
-        int Width { get; }
-        int Height { get; }
+        double Width { get; }
+        double Height { get; }
     }
 }
